Move dispatcher.js to fu-dispatcher.js
