+++ test/lib/AppDispatcherSpec.js
@@ -1,4 +1,4 @@
-import AppDispatcher from '../../src/lib/dispatcher'
+import AppDispatcher from '../../src/lib/fu-dispatcher'
 import KeyMirror from 'keymirror'
 import Invariant from 'invariant'
