+++ test/lib/specHelperSpec.js
@@ -1,7 +1,7 @@
 /* global expect */
 import ActionPrototype from '../../src/lib/ActionPrototype'
 import StorePrototype from '../../src/lib/StorePrototype'
-import Dispatcher from '../../src/lib/dispatcher'
+import Dispatcher from '../../src/lib/fu-dispatcher'
 import assign from 'object-assign'
 import KeyMirror from 'keymirror'
 import React from 'react/addons'
